refactor: optimize manage page
