+++ src/renderer/middleware/reminders.ts
@@ -3,7 +3,7 @@ import { Middleware } from '@nuxt/types'
 // Reminders are shown between pages
 const middleware: Middleware = ({ from, route, localePath, $notify }) => {
   // Reminder to fetch media when leaving the manage media page
-  if (from.path === localePath('/add') && route.path !== localePath('/add')) {
+  if (from.path === localePath('/manage') && route.path !== localePath('/manage')) {
     $notify('dontForgetToGetMedia')
   }
 }
